chore: clean up adapter after extraction
