+++ backend/internal/infra/media/ffmpeg/adapter.go
@@ -9,11 +9,9 @@ import (
 	"io"
 	"net"
 	"net/http"
-	"net/url"
 	"os"
 	"os/exec"
 	"path/filepath"
-	"regexp"
 	"strconv"
 	"strings"
 	"sync"
@@ -31,8 +29,6 @@ import (
 )
 
 const (
-	preflightMinBytes = 188 * 3
-	preflightTimeout  = 2 * time.Second
 	// safariDirtyHLSTimeSec reduces first-frame latency for dirty live sources while
 	// preserving stable 2-second GOP/segment alignment.
 	safariDirtyHLSTimeSec = 2
@@ -42,7 +38,6 @@ const (
 
 // vaapiEncodersToTest is the list of VAAPI encoders verified during preflight.
 var vaapiEncodersToTest = []string{"h264_vaapi", "hevc_vaapi"}
-var ffmpegURLPattern = regexp.MustCompile(`[A-Za-z][A-Za-z0-9+.-]*://[^'"\s]+`)
 
 // LocalAdapter implements ports.MediaPipeline using local exec.Command.
 type LocalAdapter struct {
@@ -353,17 +348,11 @@ func (a *LocalAdapter) VaapiEncoderVerified(encoder string) bool {
 
 // Start initiates the media process.
 func (a *LocalAdapter) Start(ctx context.Context, spec ports.StreamSpec) (ports.RunHandle, error) {
-	// 0. Tune if required
 	if spec.Source.Type == ports.SourceTuner && a.E2 != nil {
 		if spec.Source.TunerSlot < 0 {
 			return "", fmt.Errorf("invalid tuner slot: %d", spec.Source.TunerSlot)
 		}
-		// Create ephemeral tuner using legacy logic (reused)
-		// We use a short timeout for the tune operation itself
 		tuner := enigma2.NewTuner(a.E2, spec.Source.TunerSlot, 10*time.Second)
-
-		// Use a detached context or the start context?
-		// Start context is appropriate.
 		if err := tuner.Tune(ctx, spec.Source.ID); err != nil {
 			return "", fmt.Errorf("tuning failed: %w", err)
 		}
@@ -412,7 +401,6 @@ func (a *LocalAdapter) Start(ctx context.Context, spec ports.StreamSpec) (ports.
 	}
 	sourceKey := fpsCacheKey(spec.Source, inputURL)
 
-	// 1. Generate Arguments from Spec
 	a.Logger.Info().
 		Str("session_id", spec.SessionID).
 		Str("startup_phase", "ffmpeg_args_build_started").
@@ -428,24 +416,20 @@ func (a *LocalAdapter) Start(ctx context.Context, spec ports.StreamSpec) (ports.
 		Int("arg_count", len(args)).
 		Msg("ffmpeg args build finished")
 
-	// 2. Prepare Command
 	// #nosec G204 - BinPath is trusted from config; args are generated by strict internal logic (buildArgs)
 	cmd := exec.CommandContext(ctx, a.BinPath, args...)
 	procgroup.Set(cmd) // Mandatory for tree reaping
 
-	// 3. Setup Logging & Watchdog
 	stderr, err := cmd.StderrPipe()
 	if err != nil {
 		return "", fmt.Errorf("failed to pipe stderr: %w", err)
 	}
 	cmd.Stdout = nil
 
-	// 4. Start
 	if err := cmd.Start(); err != nil {
 		return "", fmt.Errorf("ffmpeg start failed: %w", err)
 	}
 
-	// 6. Monitor Process
 	pid := cmd.Process.Pid
 	a.Logger.Info().
 		Str("session_id", spec.SessionID).
@@ -457,15 +441,11 @@ func (a *LocalAdapter) Start(ctx context.Context, spec ports.StreamSpec) (ports.
 	a.activeProcs[handle] = cmd
 	a.mu.Unlock()
 
-	// Start watchdog monitor
 	go a.monitorProcess(ctx, handle, cmd, stderr, spec.SessionID)
 	if sourceKey != "" {
-		// Learn FPS from actual encoder output stream for future sessions of the same service_ref.
 		go a.learnFPSFromOutput(sourceKey, spec.SessionID)
 	}
 
-	// Metrics: Record pipeline spawn with cause="admitted" only AFTER successful start.
-	// We use engine="ffmpeg" as per truth.
 	metrics.RecordPipelineSpawn("ffmpeg", "admitted")
 	return handle, nil
 }
@@ -520,15 +500,11 @@ func (a *LocalAdapter) monitorProcess(parentCtx context.Context, handle ports.Ru
 		}
 		a.Logger.Error().Str("sessionId", sessionID).Str("ffmpeg_log", sanitizeFFmpegLogLine(line)).Msg("ffmpeg output")
 		wd.ParseLine(line)
-
-		// Map back to session metrics if needed, but primarily for stall detection
-
 	}
 	if scanErr := scanner.Err(); scanErr != nil {
 		a.Logger.Warn().Err(scanErr).Str("sessionId", sessionID).Msg("ffmpeg stderr scan failed")
 	}
 
-	// Wait for process or watchdog
 	procErr := cmd.Wait()
 	wdCancel()
 
@@ -536,8 +512,6 @@ func (a *LocalAdapter) monitorProcess(parentCtx context.Context, handle ports.Ru
 	case wdErr := <-wdErrCh:
 		if wdErr != nil {
 			a.Logger.Error().Err(wdErr).Str("sessionId", sessionID).Msg("watchdog triggered process termination")
-			// We don't have an easy way to signal the domain here other than the process exiting.
-			// But since we are monitoring STDEER and it closed, cmd.Wait() will return.
 		}
 	default:
 	}
@@ -640,7 +614,6 @@ func (a *LocalAdapter) Stop(ctx context.Context, handle ports.RunHandle) error {
 	return nil
 }
 
-// Health checks if the process is running.
 func (a *LocalAdapter) Health(ctx context.Context, handle ports.RunHandle) ports.HealthStatus {
 	a.mu.Lock()
 	cmd, exists := a.activeProcs[handle]
@@ -662,7 +635,6 @@ func (a *LocalAdapter) Health(ctx context.Context, handle ports.RunHandle) ports
 			LastCheck: time.Now(),
 		}
 	}
-	// Signal 0 checks liveness without sending a real signal.
 	if err := cmd.Process.Signal(syscall.Signal(0)); err != nil {
 		a.mu.Lock()
 		delete(a.activeProcs, handle)
@@ -683,95 +655,6 @@ func (a *LocalAdapter) Health(ctx context.Context, handle ports.RunHandle) ports
 		Message:   "process active",
 		LastCheck: time.Now(),
 	}
-}
-
-func hasTSSync(buf []byte) bool {
-	if len(buf) < preflightMinBytes {
-		return false
-	}
-	return buf[0] == 0x47 && buf[188] == 0x47 && buf[376] == 0x47
-}
-
-func buildFallbackURL(resolvedURL, serviceRef string) (string, error) {
-	u, err := url.Parse(resolvedURL)
-	if err != nil {
-		return "", err
-	}
-	host := u.Hostname()
-	if host == "" {
-		return "", fmt.Errorf("missing host in resolved url")
-	}
-	scheme := u.Scheme
-	if scheme == "" {
-		scheme = "http"
-	}
-	u.Scheme = scheme
-	u.Host = fmt.Sprintf("%s:%d", host, 8001)
-	u.Path = "/" + serviceRef
-	u.RawQuery = ""
-	u.Fragment = ""
-	u.User = nil
-	return u.String(), nil
-}
-
-func isStreamRelayURL(rawURL string) bool {
-	u, err := url.Parse(rawURL)
-	if err != nil {
-		return false
-	}
-	port := u.Port()
-	if port == "" {
-		port = defaultPortForScheme(u.Scheme)
-	}
-	return port == "17999"
-}
-
-func defaultPortForScheme(scheme string) string {
-	if strings.EqualFold(scheme, "https") {
-		return "443"
-	}
-	return "80"
-}
-
-func sanitizeURLForLog(rawURL string) string {
-	u, err := url.Parse(rawURL)
-	if err != nil {
-		return rawURL
-	}
-	u.User = nil
-	return u.String()
-}
-
-func sanitizeFFmpegLogLine(line string) string {
-	return ffmpegURLPattern.ReplaceAllStringFunc(line, sanitizeURLForLog)
-}
-
-func (a *LocalAdapter) injectCredentialsIfAllowed(streamURL string) string {
-	if a.E2 == nil {
-		return streamURL
-	}
-	if a.E2.Username == "" && a.E2.Password == "" {
-		return streamURL
-	}
-
-	u, err := url.Parse(streamURL)
-	if err != nil {
-		return streamURL
-	}
-
-	port := u.Port()
-	if port == "" {
-		port = defaultPortForScheme(u.Scheme)
-	}
-
-	if port == "80" || port == "443" || port == "8001" || port == "8002" {
-		if a.E2.Username != "" {
-			u.User = url.UserPassword(a.E2.Username, a.E2.Password)
-		}
-		return u.String()
-	}
-
-	return streamURL
 }
 
 func envIntBounded(key string, defaultValue, minValue, maxValue int) int {
